Zsnd: Fixed glob escape
`glob.escape()` escapes all globals, including `*` and `?`. `*` and `?` are not possible to use in Windows paths, but square brackets the only remaining glob method are, and they're not often used in glob expressions. So, Windows users are more often expecting square brackets to work in paths  than using them in glob expressions. I replaced `glob.escape()` with `.replace('[', '[[]')`, which worked in my tests. This article mentions all problems with glob, including my chosen solution: https://stackoverflow.com/questions/2186525/how-to-use-to-find-files-recursively